Fixing logic

+++ bot.js
@@ -13,7 +13,7 @@ client.on('message', message => {
         message.delete();
         message.reply('please use ' + message.guild.channels.find(channel => channel.name === "general").toString() + ' for general chat.');
     }
-    else if (message.content.startsWith('!')) {
+    else if (message.content.startsWith('!') && message.channel.name !== 'jukebox') {
         message.delete();
         message.reply('please use ' + message.guild.channels.find(channel => channel.name === "jukebox").toString() + ' to play music.');
     }
